+++ example/lib/res/images/images_keys.images_keys.dart
@@ -1,5 +1,5 @@
 // DO NOT EDIT. This is code generated via package:images_keys/generate.dart
-abstract class  ImageKeys {
-  static const logo_dart_192px_svg = 'assets/images/logo_dart_192px.svg';
+abstract class  ImagesKeys {
+  static const logo_dart_png = 'assets/images/logo_dart.png';
 
 }