+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Scans only files
+- Improve documentation
+- Add Flutter example
+
 ## 0.0.1
 
 - Initial version